Move a bunch of stuff into src/inscriptions
